refactor: Balls 생성자 인자 및 유효성 검사 추가
Balls(List<Ball> ball) -> Balls(List<Integer> number)
