+++ src/test/java/baseball/model/BallsTest.java
@@ -6,17 +6,15 @@ import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
 
+import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
+
 public class BallsTest {
 
     private Balls balls;
 
     @BeforeEach
     void setUp() {
-        balls = new Balls(Arrays.asList(
-                new Ball(0, 1),
-                new Ball(1, 2),
-                new Ball(2, 3)
-        ));
+        balls = new Balls(Arrays.asList(1, 2, 3));
     }
 
     @Test
@@ -30,5 +28,17 @@ public class BallsTest {
                 JudgeResult.BALL,
                 JudgeResult.NOTHING
         );
+    }
+
+    @Test
+    void Balls_생성시_잘못된_값_들어오면_IllegalArgumentException_에러_발생() {
+        assertThatIllegalArgumentException()
+                .isThrownBy(() -> new Balls(Arrays.asList(1, 2, 3, 4)));
+
+        assertThatIllegalArgumentException()
+                .isThrownBy(() -> new Balls(Arrays.asList(1, 2, 2)));
+
+        assertThatIllegalArgumentException()
+                .isThrownBy(() -> new Balls(Arrays.asList(0, 1, 3)));
     }
 }
